requirements.txt
asgiref==3.3.1
Django==3.0
django-recaptcha==2.0.6
djangorestframework==3.12.4
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
